refactor(land): reorganize enums
