+++ src/pland/Global.h
@@ -42,23 +42,6 @@ enum class LandPermType : int {
     Guest,        // 访客
 };
 
-/**
- * 领地持有类型
- */
-enum class LandHoldType : uint8_t {
-    Bought = 0, // 购买模式
-    Leased = 1, // 租赁模式
-};
-
-/**
- * 领地租赁状态
- */
-enum class LeaseState : uint8_t {
-    None    = 0, // 非租赁领地 / 无状态
-    Active  = 1, // 正常期
-    Frozen  = 2, // 冻结期
-    Expired = 3, // 已到期
-};
 
 inline int constexpr GlobalSubLandMaxNestedLevel = 16; // 子领地最大嵌套层数
